2019-08-26 16:57 完成仓库调拨三级审核 调整app提交单据

+++ mf-item/mf-item-interface/src/main/java/com/mallfe/item/pojo/AllBill.java
@@ -57,7 +57,7 @@ public class AllBill {
     /**
      * 数量
      */
-    private Long sl;
+    private Integer sl;
 
     private String pinm;
